7522: fix tests; some refactoring
