+++ web-client/src/presenter/computeds/menuHelper.test.js
@@ -52,7 +52,7 @@ describe('menuHelper', () => {
     const { isEditCaseTrialInformationMenuOpen } = runCompute(menuHelper, {
       state: {
         navigation: {
-          openMenu: 'EditCaseTrialInformationMenu',
+          editCaseTrialInfoMenu: 'EditCaseTrialInformationMenu',
         },
       },
     });
